refactor(utils): reorder tRPC utils folders
